core(config): split out config helpers (#9003)
